Implement bubble sort using slices and add unit tests
